added flashamr class
git-svn-id: https://svn.mcs.anl.gov/repos/osuflow/trunk@58 6b533941-e25a-0410-8d37-8675ae1e3315

+++ calc_subvolume.h
@@ -34,7 +34,7 @@ typedef
   float xmin, xmax; 
   float ymin, ymax;
   float zmin, zmax;
-  float tmin, tmax;   // computational time steps
+  int tmin, tmax;   // computational time steps
   int xdim, ydim, zdim, tdim; // data resolution 
 } volume_bounds_type_f;
